+++ include/network_manager.h
@@ -1,0 +1,13 @@
+#pragma once
+#include <Arduino.h>
+
+struct UploadParams {
+    char *runName;
+    char *track;
+    char *comments;
+};
+
+void startAP();
+void connectToWiFi();
+void setupWebRoutes();
+void uploadRunTask(void *pvParameter);
