+++ include/telemetry_tasks.h
@@ -1,0 +1,3 @@
+#pragma once
+void DataTaskcode(void * pvParameter);
+void WiFiTaskcode(void * pvParameter);
